pyproject.toml
[tool.poetry]
name = "cryptossets"
version = "0.1.0"
description = ""
authors = ["Hossein Sokhandan <iamsokhandan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.0.1"
"coinex.py" = "^1.0.4"
pandas = "^1.3.5"
numpy = "^1.22.1"
django-crispy-forms = "^1.13.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
